Extract db collation handler from Common class
