+++ libraries/classes/Common.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace PhpMyAdmin;
 
 use PhpMyAdmin\Query\Utilities;
-use function explode;
 use function strlen;
 use function strpos;
 
@@ -29,9 +28,6 @@ final class Common
             $is_show_stats = false;
         }
 
-        $relation = new Relation($dbi);
-        $operations = new Operations($dbi, $relation);
-
         /**
          * Defines the urls to return to in case of error in a sql statement
          */
@@ -47,102 +43,39 @@ final class Common
          * Ensures the database exists (else move to the "parent" script) and displays
          * headers
          */
-        if (! isset($is_db) || ! $is_db) {
-            if (strlen($db) > 0) {
-                $is_db = $dbi->selectDb($db);
-                // This "Command out of sync" 2014 error may happen, for example
-                // after calling a MySQL procedure; at this point we can't select
-                // the db but it's not necessarily wrong
-                if ($dbi->getError() && $errno == 2014) {
-                    $is_db = true;
-                    unset($errno);
-                }
-            } else {
-                $is_db = false;
-            }
-            // Not a valid db name -> back to the welcome page
-            $params = ['reload' => '1'];
-            if (isset($message)) {
-                $params['message'] = $message;
-            }
-            $uri = './index.php?route=/' . Url::getCommonRaw($params, '&');
-            if (strlen($db) === 0 || ! $is_db) {
-                if ($response->isAjax()) {
-                    $response->setRequestStatus(false);
-                    $response->addJSON(
-                        'message',
-                        Message::error(__('No databases selected.'))
-                    );
-                } else {
-                    Core::sendHeaderLocation($uri);
-                }
-                exit;
-            }
+        if (isset($is_db) && $is_db) {
+            return;
         }
 
-        /**
-         * Changes database charset if requested by the user
-         */
-        if (isset($_POST['submitcollation'], $_POST['db_collation']) && ! empty($_POST['db_collation'])) {
-            [$db_charset] = explode('_', $_POST['db_collation']);
-            $sql_query = 'ALTER DATABASE ' . Util::backquote($db)
-                . ' DEFAULT' . Util::getCharsetQueryPart($_POST['db_collation'] ?? '');
-            $dbi->query($sql_query);
-            $message = Message::success();
-
-            /**
-             * Changes tables charset if requested by the user
-             */
-            if (isset($_POST['change_all_tables_collations']) &&
-                $_POST['change_all_tables_collations'] === 'on'
-            ) {
-                [$tables] = Util::getDbInfo($db, null);
-                foreach ($tables as $tableName => $data) {
-                    if ($dbi->getTable($db, $tableName)->isView()) {
-                        // Skip views, we can not change the collation of a view.
-                        // issue #15283
-                        continue;
-                    }
-                    $sql_query = 'ALTER TABLE '
-                        . Util::backquote($db)
-                        . '.'
-                        . Util::backquote($tableName)
-                        . ' DEFAULT '
-                        . Util::getCharsetQueryPart($_POST['db_collation'] ?? '');
-                    $dbi->query($sql_query);
-
-                    /**
-                     * Changes columns charset if requested by the user
-                     */
-                    if (! isset($_POST['change_all_tables_columns_collations']) ||
-                        $_POST['change_all_tables_columns_collations'] !== 'on'
-                    ) {
-                        continue;
-                    }
-
-                    $operations->changeAllColumnsCollation($db, $tableName, $_POST['db_collation']);
-                }
+        if (strlen($db) > 0) {
+            $is_db = $dbi->selectDb($db);
+            // This "Command out of sync" 2014 error may happen, for example
+            // after calling a MySQL procedure; at this point we can't select
+            // the db but it's not necessarily wrong
+            if ($dbi->getError() && $errno == 2014) {
+                $is_db = true;
+                unset($errno);
             }
-            unset($db_charset);
-
-            /**
-             * If we are in an Ajax request, let us stop the execution here. Necessary for
-             * db charset change action on /database/operations. If this causes a bug on
-             * other pages, we might have to move this to a different location.
-             */
-            if ($response->isAjax()) {
-                $response->setRequestStatus($message->isSuccess());
-                $response->addJSON('message', $message);
-                exit;
-            }
-        } elseif (isset($_POST['submitcollation'], $_POST['db_collation']) && empty($_POST['db_collation'])) {
+        } else {
+            $is_db = false;
+        }
+        // Not a valid db name -> back to the welcome page
+        $params = ['reload' => '1'];
+        if (isset($message)) {
+            $params['message'] = $message;
+        }
+        $uri = './index.php?route=/' . Url::getCommonRaw($params, '&');
+        if (strlen($db) === 0 || ! $is_db) {
             if ($response->isAjax()) {
                 $response->setRequestStatus(false);
                 $response->addJSON(
                     'message',
-                    Message::error(__('No collation provided.'))
+                    Message::error(__('No databases selected.'))
                 );
+            } else {
+                Core::sendHeaderLocation($uri);
             }
+            exit;
         }
     }
 }
