+++ libraries/classes/Controllers/Database/OperationsController.php
@@ -344,4 +344,64 @@ class OperationsController extends AbstractController
             'collations' => $collations,
         ]);
     }
+
+    public function collation(): void
+    {
+        global $db;
+
+        if (! $this->response->isAjax()) {
+            return;
+        }
+
+        if (empty($_POST['db_collation'])) {
+            $this->response->setRequestStatus(false);
+            $this->response->addJSON('message', Message::error(__('No collation provided.')));
+
+            return;
+        }
+
+        Common::database();
+
+        $sql_query = 'ALTER DATABASE ' . Util::backquote($db)
+            . ' DEFAULT' . Util::getCharsetQueryPart($_POST['db_collation'] ?? '');
+        $this->dbi->query($sql_query);
+        $message = Message::success();
+
+        /**
+         * Changes tables charset if requested by the user
+         */
+        if (isset($_POST['change_all_tables_collations']) &&
+            $_POST['change_all_tables_collations'] === 'on'
+        ) {
+            [$tables] = Util::getDbInfo($db, null);
+            foreach ($tables as $tableName => $data) {
+                if ($this->dbi->getTable($db, $tableName)->isView()) {
+                    // Skip views, we can not change the collation of a view.
+                    // issue #15283
+                    continue;
+                }
+                $sql_query = 'ALTER TABLE '
+                    . Util::backquote($db)
+                    . '.'
+                    . Util::backquote($tableName)
+                    . ' DEFAULT '
+                    . Util::getCharsetQueryPart($_POST['db_collation'] ?? '');
+                $this->dbi->query($sql_query);
+
+                /**
+                 * Changes columns charset if requested by the user
+                 */
+                if (! isset($_POST['change_all_tables_columns_collations']) ||
+                    $_POST['change_all_tables_columns_collations'] !== 'on'
+                ) {
+                    continue;
+                }
+
+                $this->operations->changeAllColumnsCollation($db, $tableName, $_POST['db_collation']);
+            }
+        }
+
+        $this->response->setRequestStatus($message->isSuccess());
+        $this->response->addJSON('message', $message);
+    }
 }
